Fixed hard-coded paths in all the python scripts (to use  !/usr/bin/env python instead).
Added missing command-line flags: -MM, -MT, etc. to clang.py.

Attempt to fix the fixname function in par.py so that it does not
rewrite ../filename to .bc./filename (and similar nonsense). We now
explicitly check for the suffixes (.a --> .bc.a and .o --> .bc.o).

Removed explict mapping MAKE = gmake from the Makefile.