fix op_iterator bug? (not sure)
User::op_iterator cur = args_begin; is assigned once and never incremented. I believe it should be incremented at the end of inner for loop so that `i` and `cur` are tracking the same formal argument. However, I can't figure out a code in which this method would be called, so can't test it. This PR is to put it on your radar, please merge only if it makes sense.