[NEW] function pointers passed as function arguments can be specialized.
To do this we need for now to run slash with option --no-strip. The
reason is that the LLVM strip pass will get rid of all function names
if they are marked as internal. However, when specializing a pointer
to a function bar passed as argument to a function foo the function
bar can be become dead code in the current module if foo is defined in
another module and bar is only called inside bar. This means that bar
was initially internal but then during the specialization becomes
external. Rather than disabling strip possibly a better solution is to
strip symbols once the inter-specialization took place.