Relax conditions when a function/global should be internalized.
The semantics of linkage types in LLVM is not formally defined so it's
hard to figure out what the linker actually does. Most of the
confusion comes from the different variants of weak semantics.

My take for now is to convert almost any linkage type to internal
linkage if we know that the symbol cannot be used in other compilation
units. The other change I made is that I treat functions and globals
in the same way.