WIP: Interpreter for OCCAM programs.
The interpreter is right now a direct copy of the LLVM interpreter.

We have made a copy because we need to modify it in several
ways. First, we want to allow the interpreter to have unknown values
as long as it's not used in a branch. Second, we want to stop as soon
as we cannot tell the evaluation of branch if it depends on some
unknown value.