WIP on how to use the partial information collected by the interpreter.
We need to identify a cutpoint (some basic block) on which we can
claim that that information collected from the interpreter is
sound. This will be an assumption because the user can always pass
parameters at runtime that overwrite everything.

Then, we need to decide how to use the information at the
cutpoint. This is tricky if a variable is a pointer (which most,
included global variables, are). If we know that the content a pointer
p is 2 then we would like to write at the cutpoint assume(*p ==
2). Unfortunately, LLVM passes won't understand that. Instead, we need
to identify p's reads and do the instantiation.