The config prime stuff has been hooked up to slash.
Once the interpreter is stopped, we only consider global variables
that contain basic types. Moreover, we only replace uses in the same
function where the execution stopped. This is very limited but it is
enough for programs such as tree.

More tested is needed and more work to go beyond simple global
variables and replace uses across different functions.