fix: callAny --> callFrom and replaced with empty implementation.
This is an important change that we can revert if needed.

This has to do with whether an interface should record both incoming
and outgoing calls of a module or only outgoing calls.

callTo(f, ...) records the outgoing calls of a module. That is,
external calls done by the module. callAny(f) seems to be called only
in GatherInterface when an entry point of a module is detected (by
having an edge from getExternalCallingNode to f in the
CallGraph). This is the reason to rename it to something more
informative like callFrom since it records the calls from other
modules.

More importantly, it is very confusing the fact that callAny(f) stores
a call to f with unknown parameters in the same internal datastructure
that callTo does __within__ the same interface. It would make more
sense to do callAny(f) from the interfaces of the __other__ modules to
reflect that those modules can call f. After more thinking, I believe
that we do not need to keep track if a function can be called from
other modules within the interface. It should be fine to know whether
the address of a function can be taken or not and this information is
already available.

For these reasons, I decided to keep the method callFrom but with an
empty implementation to remind me for now these issues.