refactor(razor): disable code that runs seahorn guided by ROP metrics.
I moved most of the code to a new file called rop_guided_dce.py.

I do not want to maintain code that nobody uses.  If you are really
interested in this feature and have an use case then let me know and I
will be happy to help to enable again the code.