fix(interfaces): check to add a new call
Given a new call we check first if the interface already contains that
call. For that, the interface does not care about the actual LLVM
values of the callsite parameters but instead an abstraction of it
given by InterfaceTypes.

The new check is: add a new call in the interface unless there is an
existing call with the __same__ arguments (modulo type
abstraction). This means that if there is already an entry with
either: (1) different number of arguments, or (2) types do not match,
or (3) types match but the new call has more refined types then we add
the new call in the interface.