perf(py): Compute initial interfaces without seadsa callgraphs
We compute the initial interfaces using LLVM callgraphs.

After intra- and inter-specialization and optimization have been
completed we use seadsa callgraphs to compute again interfaces and
propagate informatino until fixpoint.

This approach should not affect precision but it should be faster.
The idea is to use a cheap callgraph first and then use a more
expensive but precise callgraph once the modules have been already
optimized.