fix: communication between llvm and python.
The python part knows that some llvm pass did progress by writing to
the standard error some message like .... progress ....  This is so
hacky but it is what it is. I added those messages in the LLVM passes
called by python previrt_progress. The original code was probably
correct but after tons of cleanups I probably introduced the bugs
since there was no coment or documentation that could tell those
messages are vital.