feat: specialize libraries.
I merged manually the PR #29.

This PR allows to specialize libraries with respect to main and also a
wrt to set of main functions. It adds two manifest keys:

- lib_spec: libraries to be specialized.
- main_spec: main functions used to specialize libraries in lib_spec.

In addition, there is a new slash option --entry-points that allows
users to tell the entry points of a library. Otherwise all
non-internal functions are considered as entry points.

I have not tested this new functionality but better to merge it before
forget about it.

fix: script to run liboccamize test

fix: string.h --> string

fix: link with LLVMDemangle

test: update tests for gclang and new manifest