refactor(configPrime): if FFI only main argv is not tracked.
We kept track of all the memory allocated by the LLVM module. In that
way, if we read from some unallocated memory then we can stop
gracefully the LLVM interpreter. However, this approach is too
conservative if an external call allocates memory.

Thus, if FFI is enabled then we assume that all memory is allocated
except those originated from main argv.