fix(config-prime): do not abort if global cannot be resolved
We relied on LLVM emitGlobals to allocate memory for globals.  If a
global cannot be resolved then the code reports an error. We do not
want to crash here. Instead, we ignore the global and keep track of it
so that when we try to read from it an unknown value is returned.

This happens with wget with the global gnutls_free.