feat: use seadsa as alias analysis in llvm opt.
This allows llvm opt to leverage seadsa anytime.

I was not able to plugin sea-dsa as alias analysis in vanilla llvm
opt. My solution is to have our own version of llvm opt and modify
the code to plugin seadsa there.

Rather than making a copy of llvm 10 opt, SeaHorn already has its own
version of llvm opt (https://github.com/seahorn/llvm-seahorn). We have
created a branch in this repo (dev10-seadsa) and include it in OCCAM as a
submodule. The advantage of doing this is that OCCAM does not need to
port to new LLVM versions as long as it is using the same version
than SeaHorn.

The llvm opt command line option to use seadsa is
--seaopt-use-seadsa-aa.

The slash command line option is --use-seaopt.