fix(config-priming): disable simplifySpeculativeSuffixes
The intention is to use information from the partial execution (called
prefix) to simplify the rest of other paths (called suffixes). This
simplification cannot be sound without pointer analysis. I prefer not
to run for now the pointer analysis because it can be expensive at
this stage and the benefits are not clear. Thus, we disable this for
now.