fix(config-prime): understanding complex semantics of getopt
If the environment variable POSIXLY_CORRECT is not set then getopt
will permute the argv’s elements so that all the ones started with '-'
come first but only if you use the glibc version.  That is, if you
call directly getopt directly (e.g., if the program is linked with
musllvm) then getopt does not permute elements.  The permutation must
be disabled because it will go through the dynamic arguments and check
if it starts with . However, the dynamic arguments cannot be
accessed during the configuration priming because they are undefined
at that time.

More details can be found here:

    https://code.woboq.org/userspace/glibc/posix/getopt.c.html#58