fix(config-prime): avoid load optarg if dynamic arg
The interpreter is taught to avoid accessing any memory that contains
dynamic arguments.  However, when using getopt via FFI we don't have
control about what getopt can do. In particular, getopt and friends
can load a dynamic argument in optarg and then the program can access
to it. One drastic solution would be to avoid also read from
optarg. Instead, we add a LLVM transformation that allow loading
optarg only if we know that it cannot contain a dynamic argument. We
can know that because we can ask optind. After the interpreter
finishes we undo the program transformation.