fix(config-prime): do not overspecialize in the interpreter
This is a fix over commit a34876418efcc2d5f7f92cd44520a8001ca4117e.

That commit did not specialize a value if it was the returned value of
some blacklisted function.  This was done after the interpreter
finished during the simplification of the trace prefix.  However, we
need to do it during the execution of the interpreter (by marking as
unknown the returned value) because blacklisted returned values can be
stored to memory and then it's too hard to identify later.